extracted text-overlays into its own file
